backend/imgtopdf/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
img2pdf==0.5.1
Pillow==11.0.0
gunicorn==21.2.0
gevent==24.2.1
